switch test to oslib for better understanding
